Update Project.ipynb
Hello, I have made some modifications to the code. I think we now have usable vectors. Although the preprocessing of images is still very simple. But I've been able to reduce the size of the images and turn them into grayscale.
At the end of the program, I also included a very simple version of a neural network, just to test the preprocessing of the images.